+++ modules/genealogy-collaboration-filament/src/Resources/CollaborationDiscussionResource/Pages/EditCollaborationDiscussion.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Liberu\Genealogy\Collaboration\Filament\Resources\CollaborationDiscussionResource\Pages;
+
+use Filament\Resources\Pages\EditRecord;
+use Liberu\Genealogy\Collaboration\Filament\Resources\CollaborationDiscussionResource;
+
+final class EditCollaborationDiscussion extends EditRecord
+{
+    protected static string $resource = CollaborationDiscussionResource::class;
+}